// src/application/invoices/validation/models/invoice-validation-error-code.ts
export enum InvoiceValidationErrorCode {
  MISSING_REQUEST = "MISSING_REQUEST",
  MISSING_INVOICE_IDENTIFIER = "MISSING_INVOICE_IDENTIFIER",
  MISSING_RESERVATION_REFERENCE = "MISSING_RESERVATION_REFERENCE",
  MISSING_CUSTOMER_REFERENCE = "MISSING_CUSTOMER_REFERENCE",
  MISSING_QUOTE_REFERENCE = "MISSING_QUOTE_REFERENCE",
  MISSING_PRICING_SNAPSHOT = "MISSING_PRICING_SNAPSHOT",
  MISSING_FINANCIAL_OBLIGATION = "MISSING_FINANCIAL_OBLIGATION",
  INVALID_CURRENCY = "INVALID_CURRENCY",
  INVALID_TOTAL_AMOUNT = "INVALID_TOTAL_AMOUNT",
  MISSING_RESERVATION = "MISSING_RESERVATION",
  RESERVATION_NOT_CONFIRMED = "RESERVATION_NOT_CONFIRMED",
  RESERVATION_CANCELLED = "RESERVATION_CANCELLED",
  QUOTE_REFERENCE_INCONSISTENT = "QUOTE_REFERENCE_INCONSISTENT",
  PRICING_REFERENCE_INCONSISTENT = "PRICING_REFERENCE_INCONSISTENT",
  PRICING_CURRENCY_MISMATCH = "PRICING_CURRENCY_MISMATCH",
  PRICING_TOTAL_MISMATCH = "PRICING_TOTAL_MISMATCH",
  INVALID_DEPOSIT_REQUIREMENT = "INVALID_DEPOSIT_REQUIREMENT",
  INVALID_PAYMENT_ALLOCATION = "INVALID_PAYMENT_ALLOCATION",
  PAYMENT_ALLOCATION_TOTAL_MISMATCH = "PAYMENT_ALLOCATION_TOTAL_MISMATCH",
  INVALID_AMOUNT_PAID = "INVALID_AMOUNT_PAID",
  INVALID_BALANCE_DUE = "INVALID_BALANCE_DUE",
  INVALID_REFUNDABLE_AMOUNT = "INVALID_REFUNDABLE_AMOUNT",
  INVALID_ADJUSTMENT = "INVALID_ADJUSTMENT",
  INVALID_CANCELLATION_SNAPSHOT = "INVALID_CANCELLATION_SNAPSHOT",
  INVALID_DUE_DATE = "INVALID_DUE_DATE",
  INVALID_INVOICE_STATUS = "INVALID_INVOICE_STATUS",
  INVOICE_NOT_READY = "INVOICE_NOT_READY",
  INVOICE_ALREADY_VOID = "INVOICE_ALREADY_VOID",
  INVOICE_ALREADY_CANCELLED = "INVOICE_ALREADY_CANCELLED",
  INVALID_EXTERNAL_REFERENCE = "INVALID_EXTERNAL_REFERENCE",
  INVALID_METADATA = "INVALID_METADATA",
}